refactor of the save and continue commands
