+++ src/Elsa.Server.Tests/Features/Workflow/LoadWorkflowActivity/LoadWorkflowActivityRequestHandlerTests.cs
@@ -1,5 +1,4 @@
 ﻿using AutoFixture.Xunit2;
-using Elsa.CustomActivities.Activities.MultipleChoice;
 using Elsa.CustomActivities.Activities.Shared;
 using Elsa.CustomInfrastructure.Data.Repository;
 using Elsa.CustomModels;
@@ -37,7 +36,7 @@ namespace Elsa.Server.Tests.Features.Workflow.LoadWorkflowActivity
 
             //Arrange
             multipleChoiceQuestionInvoker
-                .Setup(x => x.FindWorkflowsAsync<MultipleChoiceQuestion>(loadWorkflowActivityRequest.ActivityId, loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
+                .Setup(x => x.FindWorkflowsAsync(loadWorkflowActivityRequest.ActivityId, "MultipleChoiceQuestion", loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
                 .ReturnsAsync(collectedWorkflows);
 
             workflowInstanceStore.Setup(x =>
@@ -91,7 +90,7 @@ namespace Elsa.Server.Tests.Features.Workflow.LoadWorkflowActivity
 
             //Arrange
             multipleChoiceQuestionInvoker
-                .Setup(x => x.FindWorkflowsAsync<MultipleChoiceQuestion>(loadWorkflowActivityRequest.ActivityId, loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
+                .Setup(x => x.FindWorkflowsAsync(loadWorkflowActivityRequest.ActivityId, "MultipleChoiceQuestion", loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
                 .ReturnsAsync(collectedWorkflows);
 
             workflowInstanceStore.Setup(x =>
@@ -143,7 +142,7 @@ namespace Elsa.Server.Tests.Features.Workflow.LoadWorkflowActivity
 
             //Arrange
             multipleChoiceQuestionInvoker
-                .Setup(x => x.FindWorkflowsAsync<MultipleChoiceQuestion>(loadWorkflowActivityRequest.ActivityId, loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
+                .Setup(x => x.FindWorkflowsAsync(loadWorkflowActivityRequest.ActivityId, "MultipleChoiceQuestion", loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
                 .ReturnsAsync(collectedWorkflows);
 
             workflowInstanceStore.Setup(x =>
@@ -180,7 +179,7 @@ namespace Elsa.Server.Tests.Features.Workflow.LoadWorkflowActivity
         {
             //Arrange
             multipleChoiceQuestionInvoker
-                .Setup(x => x.FindWorkflowsAsync<MultipleChoiceQuestion>(loadWorkflowActivityRequest.ActivityId, loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
+                .Setup(x => x.FindWorkflowsAsync(loadWorkflowActivityRequest.ActivityId, "MultipleChoiceQuestion", loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
                 .ReturnsAsync(collectedWorkflows);
 
             workflowInstanceStore.Setup(x =>
@@ -216,7 +215,7 @@ namespace Elsa.Server.Tests.Features.Workflow.LoadWorkflowActivity
         {
             //Arrange
             multipleChoiceQuestionInvoker
-                .Setup(x => x.FindWorkflowsAsync<MultipleChoiceQuestion>(loadWorkflowActivityRequest.ActivityId, loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
+                .Setup(x => x.FindWorkflowsAsync(loadWorkflowActivityRequest.ActivityId, "MultipleChoiceQuestion", loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
                 .ReturnsAsync(collectedWorkflows);
 
             workflowInstanceStore.Setup(x =>
@@ -247,7 +246,7 @@ namespace Elsa.Server.Tests.Features.Workflow.LoadWorkflowActivity
         {
             //Arrange
             multipleChoiceQuestionInvoker
-                .Setup(x => x.FindWorkflowsAsync<MultipleChoiceQuestion>(loadWorkflowActivityRequest.ActivityId, loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
+                .Setup(x => x.FindWorkflowsAsync(loadWorkflowActivityRequest.ActivityId, "MultipleChoiceQuestion", loadWorkflowActivityRequest.WorkflowInstanceId, CancellationToken.None))
                 .ReturnsAsync(new List<CollectedWorkflow>());
 
             //Act
@@ -275,7 +274,7 @@ namespace Elsa.Server.Tests.Features.Workflow.LoadWorkflowActivity
         {
             //Arrange
             multipleChoiceQuestionInvoker
-                .Setup(x => x.FindWorkflowsAsync<MultipleChoiceQuestion>(It.IsAny<string>(), It.IsAny<string>(), CancellationToken.None))
+                .Setup(x => x.FindWorkflowsAsync(It.IsAny<string>(), "MultipleChoiceQuestion", It.IsAny<string>(), CancellationToken.None))
                 .Throws(exception);
 
             //Act
