+++ src/Elsa.Server/Features/Currency/SaveAndContinue/SaveAndContinueCommand.cs
@@ -1,17 +1,14 @@
-﻿using Elsa.Server.Models;
+﻿using Elsa.Server.Features.Shared;
+using Elsa.Server.Features.Shared.SaveAndContinue;
+using Elsa.Server.Models;
 using MediatR;
 
 namespace Elsa.Server.Features.Currency.SaveAndContinue
 {
 
 
-    public class SaveAndContinueCommand : IRequest<OperationResult<SaveAndContinueResponse>>
+    public class SaveAndContinueCommand : SaveAndContinueCommandBase, IRequest<OperationResult<SaveAndContinueResponse>>
     {
-        public string Id { get; set; } = null!;
-        public string ActivityId { get; set; } = null!;
-
-        public string WorkflowInstanceId { get; set; } = null!;
-
         public string? Answer { get; set; }
     }
 }
