+++ src/Elsa.Server/Features/Currency/SaveAndContinue/SaveAndContinueMapper.cs
@@ -1,36 +1,36 @@
-﻿using Elsa.CustomModels;
-using Elsa.Server.Providers;
+﻿//using Elsa.CustomModels;
+//using Elsa.Server.Providers;
 
-namespace Elsa.Server.Features.Currency.SaveAndContinue
-{
-    public interface ISaveAndContinueMapper
-    {
-        MultipleChoiceQuestionModel SaveAndContinueCommandToNextMultipleChoiceQuestionModel(SaveAndContinueCommand command, string nextActivityId, string nextActivityType);
-    }
+//namespace Elsa.Server.Features.Currency.SaveAndContinue
+//{
+//    public interface ISaveAndContinueMapper
+//    {
+//        MultipleChoiceQuestionModel SaveAndContinueCommandToNextMultipleChoiceQuestionModel(SaveAndContinueCommand command, string nextActivityId, string nextActivityType);
+//    }
 
-    public class SaveAndContinueMapper : ISaveAndContinueMapper
-    {
-        private readonly IDateTimeProvider _dateTimeProvider;
+//    public class SaveAndContinueMapper : ISaveAndContinueMapper
+//    {
+//        private readonly IDateTimeProvider _dateTimeProvider;
 
-        public SaveAndContinueMapper(IDateTimeProvider dateTimeProvider)
-        {
-            _dateTimeProvider = dateTimeProvider;
-        }
+//        public SaveAndContinueMapper(IDateTimeProvider dateTimeProvider)
+//        {
+//            _dateTimeProvider = dateTimeProvider;
+//        }
 
-        public MultipleChoiceQuestionModel SaveAndContinueCommandToNextMultipleChoiceQuestionModel(SaveAndContinueCommand command, string nextActivityId, string nextActivityType)
-        {
-            return new MultipleChoiceQuestionModel
-            {
-                Id = $"{command.WorkflowInstanceId}-{nextActivityId}",
-                ActivityId = nextActivityId,
-                ActivityType = nextActivityType,
-                FinishWorkflow = false,
-                NavigateBack = false,
-                Answer = null,
-                WorkflowInstanceId = command.WorkflowInstanceId,
-                PreviousActivityId = command.ActivityId,
-                CreatedDateTime = _dateTimeProvider.UtcNow()
-            };
-        }
-    }
-}
+//        public MultipleChoiceQuestionModel SaveAndContinueCommandToNextMultipleChoiceQuestionModel(SaveAndContinueCommand command, string nextActivityId, string nextActivityType)
+//        {
+//            return new MultipleChoiceQuestionModel
+//            {
+//                Id = $"{command.WorkflowInstanceId}-{nextActivityId}",
+//                ActivityId = nextActivityId,
+//                ActivityType = nextActivityType,
+//                FinishWorkflow = false,
+//                NavigateBack = false,
+//                Answer = null,
+//                WorkflowInstanceId = command.WorkflowInstanceId,
+//                PreviousActivityId = command.ActivityId,
+//                CreatedDateTime = _dateTimeProvider.UtcNow()
+//            };
+//        }
+//    }
+//}
