+++ src/Elsa.Server/Features/MultipleChoice/SaveAndContinue/SaveAndContinueCommand.cs
@@ -1,15 +1,12 @@
-﻿using Elsa.Server.Models;
+﻿using Elsa.Server.Features.Shared;
+using Elsa.Server.Features.Shared.SaveAndContinue;
+using Elsa.Server.Models;
 using MediatR;
 
 namespace Elsa.Server.Features.MultipleChoice.SaveAndContinue
 {
-    public class SaveAndContinueCommand : IRequest<OperationResult<SaveAndContinueResponse>>
+    public class SaveAndContinueCommand : SaveAndContinueCommandBase, IRequest<OperationResult<SaveAndContinueResponse>>
     {
-        public string Id { get; set; } = null!;
-        public string ActivityId { get; set; } = null!;
-
-        public string WorkflowInstanceId { get; set; } = null!;
-
         public List<string> Answers { get; set; } = null!;
     }
 }
