+++ src/Elsa.Server/Features/MultipleChoice/SaveAndContinue/SaveAndContinueCommandHandler.cs
@@ -3,6 +3,7 @@ using Elsa.CustomInfrastructure.Data.Repository;
 using Elsa.Models;
 using Elsa.Persistence;
 using Elsa.Persistence.Specifications.WorkflowInstances;
+using Elsa.Server.Features.Shared.SaveAndContinue;
 using Elsa.Server.Models;
 using Elsa.Server.Providers;
 using Elsa.Services;
