+++ src/Elsa.Server/Features/MultipleChoice/SaveAndContinue/SaveAndContinueResponse.cs
@@ -1,8 +1,8 @@
-﻿namespace Elsa.Server.Features.MultipleChoice.SaveAndContinue
-{
-    public class SaveAndContinueResponse
-    {
-        public string WorkflowInstanceId { get; set; } = null!;
-        public string NextActivityId { get; set; } = null!;
-    }
-}
+﻿//namespace Elsa.Server.Features.MultipleChoice.SaveAndContinue
+//{
+//    public class SaveAndContinueResponse
+//    {
+//        public string WorkflowInstanceId { get; set; } = null!;
+//        public string NextActivityId { get; set; } = null!;
+//    }
+//}
