+++ src/Elsa.Server/Features/Shared/SaveAndContinueResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace Elsa.Server.Features.Shared.SaveAndContinue
+{
+    public class SaveAndContinueResponse
+    {
+        public string WorkflowInstanceId { get; set; } = null!;
+        public string NextActivityId { get; set; } = null!;
+    }
+}
