+++ src/Elsa.Server/Program.cs
@@ -7,7 +7,7 @@ using Elsa.CustomInfrastructure.Data.Repository;
 using Elsa.Persistence.EntityFramework.Core.Extensions;
 using Elsa.Persistence.EntityFramework.Sqlite;
 using Elsa.Runtime;
-using Elsa.Server.Features.Currency.SaveAndContinue;
+using Elsa.Server.Features.Shared.SaveAndContinue;
 using Elsa.Server.Features.Workflow.LoadWorkflowActivity;
 using Elsa.Server.Features.Workflow.StartWorkflow;
 using Elsa.Server.Providers;
@@ -54,7 +54,6 @@ builder.Services.AddMediatR(typeof(Program).Assembly);
 builder.Services.AddScoped<IDateTimeProvider, DateTimeProvider>();
 
 builder.Services.AddScoped<IStartWorkflowMapper, StartWorkflowMapper>();
-builder.Services.AddScoped<Elsa.Server.Features.MultipleChoice.SaveAndContinue.ISaveAndContinueMapper, Elsa.Server.Features.MultipleChoice.SaveAndContinue.SaveAndContinueMapper>();
 builder.Services.AddScoped<ISaveAndContinueMapper, SaveAndContinueMapper>();
 
 builder.Services.AddScoped<ILoadWorkflowActivityJsonHelper, LoadWorkflowActivityJsonHelper>();
